+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Added `drawsc` library featuring scalable draw functions
 - Updated Simplified Chinese localization (by @TheOnly8Z)
 - Updated Italian localization (by @ThePlatynumGhost)
+- Added a hook 'TTTModifyObservingEntity' that is called before the targetid gets drawn
 
 ### Changed